GetPartFromGroup() returns all possible hits.
This starts the work on returning multiple hits in lookup function where
it makes sense.

That behavior is necessary for things like data objects that are signed
by multiple entities. At the moment, verify could fail if the first
signature block found in the SIF file is not the one we have the key
for.

This will also be useful for supporting multiple partitions in one SIF
file but also anytime multiple objects of the same kind have to be
manipulated.

Signed-off-by: Yannick Cote <yhcote@gmail.com>