Revert HdrVersion to "01"
In f68560db HdrVersion was changed to "02" to account for the
introduction of the "data generic" datatype.

Since the binary layout of the header is not actually changing, only the
value stored in one of the fields, revert HdrVersion to "01". Tools like
singularity and siftool will simply report an unknown datatype if they
run into an image with this type of content. They will still be able to
deal with it (at least at a high level).

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>