Handle C strings more consistently
The SIF binary contains C strings (bytes terminated by \000) and there
are not handled consistently when they are output.

Since there's nothing in the actual code that is doing this (other than
having an array initialized to zeroes and not writing anything to the
last element), be a little bit lenient when loading data from the file:
examine the last byte and if it's \000, ignore it.

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>