Use io.Reader instead of os.File in DescriptorInput
In order to improve testability, change the Fp field of DescriptorInput
from *os.File to io.Reader (since io.Reader is the only thing the
underlying code needs to access the data).

Modify the tests so that they can open a file to obtain the data and
pass it as a io.Reader in DescriptorInput.

This allows tests to pass other types implementing an io.Reader to avoid
hitting the filesystem.

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>