Exercise "pipe" path in AddObject
There's a path in the AddObject function that depends on reading a
"pipe" (defined by having a file handle but zero size for the input).

Implement a test for AddObject that goes thru that path.

Fix SetName and SetExtra, which fail to fill the rest of the respective
structures with zeroes when their inputs are too short.

Also fix mockSifReadWriter to properly set the lenght of the underlying
bytes buffer.

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>