Add script to verify that vendor dir is up-to-date
Add scripts/check-vendor to check that the vendor/ directory is
up-to-date wrt go.mod and that the vendor/ directory has not been
modified wrt to what should be there.

Add call to scripts/check-vendor in .circleci/config.yml so that we
verify that vendor/ and/or go.mod has not been updated without updating
the other, too.

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>