feat: add GetHeaderIntegrityReader
In support of removing direct usage of (*FileImage).Header,
GetHeaderIntegrityReader returns an io.Reader that reads the
integrity-protected fields of a given SIF image header.

Relevant test code and supporting golden files have been moved to
reflect this change. The golden file contents have not been modified,
confirming that this new implementation is equivalent to the previous
implementation.